requirements.txt
flask==3.0.0
pandas==2.2.0
matplotlib==3.8.2
python-pptx==0.6.23